Removed extra operations from JoinBuilder
